ChoiceTable factored out into a live component
There are still issues with Choice Edits that need to be fixed.  Choice Edits currently create a temporary duplicate on clients that did not generate the update.  If clients press refresh, their state will correctly remove the old version of the updated choice.  Until they press refresh, both a new and an old version of the Edited/updated choice will remain on their client view.