Add random button with pubsub
This button generates a random number that will broadcast to all clients connected to server via pubsub.

This is intended to eventually select a random choice, but will start off by generating a randon number as proof of concept.  The elements in index.ex should be factored out into a separate random.ex file